Update modular inverse using Euclidean Division.
The current explanation only focuses on its usage for precomputing inverse for every number within [1, m-1] but this technique can also be efficiently (and most importantly, in a very simple manner) used to compute the inverse for a single number. Therefore, I want to update the doc to emphasize it better.