Update kuhn_maximum_bipartite_matching.md
Two important terms are defined: maximal and maximum matching. The formulation of the Beurge's lemma requires maximum matching term definition. And it is important to show difference between maximal and maximum matching. It is also possible to state perfect and near perfect matching definition, but this particular article doesn't require them.